Database fixes in actions file and reupload of model overcast-body
Fixed the model, used the model that was being used in the docker image as well. Fixed the height and weight datatypes as well so that they are stored in required format in DB. Added dockerfile too. Moved the unused model to models archieve. Updated the action server name in endpoints as well.